Fixing small errores with LatencyTest

+++ utils/pcTests/LatencyTest/LatencyTestTypes.cpp
@@ -20,7 +20,7 @@ bool LatencyDataType::serialize(void*data,SerializedPayload_t* payload)
 	*(uint32_t*)payload->data = lt->seqnum;
 	*(uint32_t*)(payload->data+4) = (uint32_t)lt->data.size();
 	std::copy(lt->data.begin(),lt->data.end(),payload->data+8);
-	payload->length = 8+lt->data.size();
+	payload->length = (uint16_t)(8+lt->data.size());
 	return true;
 }
